pyproject.toml
[project]
name = "todo-files"
version = "0.1.0"
description = "File microservice for TODO application"
authors = [
    { name = "Kirill", email = "kirill.samal@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.32.0,<1.0.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "asyncpg (>=0.30.0,<1.0.0)",
    "psycopg (>=3.2.0,<4.0.0)",
    "alembic (>=1.13.0,<2.0.0)",
    "pydantic (>=2.9.0,<3.0.0)",
    "pydantic-settings (>=2.5.0,<3.0.0)",
    "aioboto3 (>=13.0.0,<14.0.0)",
    "aiokafka (>=0.11.0,<1.0.0)",
    "pyjwt (>=2.9.0,<3.0.0)",
    "httpx (>=0.27.0,<1.0.0)",
    "python-multipart (>=0.0.12,<1.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.3.0,<9.0.0)",
    "pytest-asyncio (>=0.24.0,<1.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "pylint (>=4.0.2,<5.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "types-redis (>=4.6.0.20241004,<5.0.0.0)",
]

[tool.black]
line-length = 88
target-version = ["py311", "py312", "py313"]

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["app"]

[tool.mypy]
ignore_missing_imports = true
explicit_package_bases = true
exclude = "migrations/.*"

[tool.pylint.main]
ignore = [".git", "__pycache__", ".venv", "build", "dist"]
ignore-paths = ["^migrations/"]

[tool.pylint.messages_control]
disable = [
    "unused-argument",
    "no-member",
    "too-few-public-methods",
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "import-outside-toplevel",
    "invalid-name",
    "too-many-ancestors",
    "duplicate-code",
    "too-many-locals",
    "import-error",
    "too-many-positional-arguments",
    "too-many-arguments",
    "too-many-return-statements",
    "assignment-from-no-return",
    "not-callable",
    "mixed-line-endings",
]

[tool.pylint.format]
max-line-length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = [
    "E203",
    "W503",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
]
exclude = [".git", "__pycache__", ".venv", "build", "dist", "*.egg-info", "migrations"]
per-file-ignores = ["app/__init__.py:F401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
